+++ src/network/ssl/socket.h
@@ -1,4 +1,4 @@
-/* $Id: socket.h,v 1.20 2005/04/12 21:49:09 jonas Exp $ */
+/* $Id: socket.h,v 1.21 2005/04/27 15:15:01 jonas Exp $ */
 
 #ifndef EL__SSL_CONNECT_H
 #define EL__SSL_CONNECT_H
@@ -8,8 +8,8 @@
 struct socket;
 
 int ssl_connect(struct socket *socket);
-int ssl_write(struct socket *socket, unsigned char *data, int len);
-int ssl_read(struct socket *socket, unsigned char *data, int len);
+ssize_t ssl_write(struct socket *socket, unsigned char *data, int len);
+ssize_t ssl_read(struct socket *socket, unsigned char *data, int len);
 int ssl_close(struct socket *socket);
 
 #endif
